@@ -1,4 +1,6 @@
-﻿using NSubstitute;
+﻿using System.Threading.Tasks;
+
+using NSubstitute;
 using Xunit;
 using McMaster.Extensions.CommandLineUtils;
 
